requirements.txt
blinker
certifi
charset-normalizer
click
colorama
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
itsdangerous
Jinja2
MarkupSafe
pillow
requests
SQLAlchemy
typing_extensions
urllib3
Werkzeug
WTForms